requirements.txt
anytree==2.8.0
apparmor==3.0.7
appdirs==1.4.4
application-utility==1.3.2
argon2==0.1.10
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
Beaker==1.11.0
btrfsutil==5.18.1
CacheControl==0.12.6
catfish==4.16.4
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
colorama==0.4.5
commonmark==0.9.1
configobj==5.1.0.dev0
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.5
distro==1.7.0
Django==4.0.6
django-browser-reload==1.6.0
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.1.5
django-filter==22.1
django-import-export==2.8.0
django-js-asset==2.0.0
django-tailwind==3.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
docopt==0.6.2
docutils==0.19
drf-yasg==1.21.3
env==0.1.0
et-xmlfile==1.1.0
evdev==1.6.0
future==0.18.2
gufw==22.4.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
inflection==0.5.1
itypes==1.2.0
jaraco.context==4.1.1
jaraco.functools==3.5.1
jaraco.text==3.8.1
Jinja2==3.1.2
keyutils==0.6
lensfun==0.3.3
LibAppArmor==3.0.7
lightdm-gtk-greeter-settings==1.2.2
lit==14.0.6.dev0
louis==3.22.0
lxml==4.9.1
Mako==1.1.6
manjaro-sdk==0.1
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
menulibre==2.2.3
more-itertools==8.13.0
msgpack==1.0.3
mugshot==0.4.3
netsnmp-python==1.0a1
npyscreen==4.10.5
nspektr==0.4.0
odfpy==1.4.1
openpyxl==3.0.10
ordered-set==4.0.2
packaging==21.3
pacman-mirrors==4.23.2
pep517==0.12.0
pexpect==4.8.0
Pillow==9.2.0
platformdirs==2.5.2
ply==3.11
progress==1.6
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.7.0
pycairo==1.21.0
pycparser==2.21
Pygments==2.12.0
PyGObject==3.42.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySide6==6.3.1
python-distutils-extra==2.39
pytz==2022.1
pyxdg==0.28
PyYAML==6.0
rados==2.0.0
rbd==2.0.0
requests==2.28.1
resolvelib==0.8.1
retrying==1.3.3
rgw==2.0.0
rich @ file:///build/python-rich/src/rich-12.5.1/dist/rich-12.5.0-py3-none-any.whl
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
shiboken2==5.15.5
shiboken6==6.3.1
six==1.16.0
sqlparse==0.4.2
systemd-python==234
tablib==3.2.1
TBB==0.2
team==1.0
tenacity==8.0.1
terminator==2.1.1
toml==0.10.2
tomli==2.0.1
trove-classifiers==2022.8.7
typing_extensions==4.3.0
udiskie==2.4.2
ufw==0.36.1
uritemplate==4.1.1
urllib3==1.26.11
validate==5.1.0.dev0
validate-pyproject==0.9
vboxapi==1.0
webencodings==0.5.1
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
